pyproject.toml
[tool.poetry]
name = "triagebot"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
fastapi = "^0.110.0"
uvicorn = "^0.18.2"
sqlalchemy = "^1.4.36"
psycopg2-binary = "^2.9.3"
requests = "^2.27.1"
pydantic = "^2.4.2"
httpx = "^0.27.0"
pytest = "^7.1.2"
litellm = "^1.36.17"
langchain-openai = "^0.1.8"
crewai = ">=0.114.0,<0.115.0"
python-jose = "^3.4.0"
passlib = "^1.7.4"
python-multipart = "^0.0.20"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
